feat: [EPMRPP-115662] add highlight to search on test case library
