+++ app/src/common/utils/index.js
@@ -73,3 +73,4 @@ export { commonValidators } from './validation';
 export { copyToClipboard } from './clipboard';
 export { getRouterParams } from './getRouterParams';
 export { convertBytesToMB } from './convertBytesToMB';
+export { highlightText } from './textHighlight';
