+++ app/src/pages/inside/common/testCaseList/testCaseNameCell/testCaseNameCell.scss
@@ -66,3 +66,7 @@
   flex-wrap: wrap;
   gap: 4px;
 }
+
+.highlight {
+  color: var(--rp-ui-color-primary-text);
+}
